Move test type to test source set
