+++ mosaic-terminal/src/commonTest/kotlin/com/jakewharton/mosaic/terminal/PlatformInputWriter.kt
@@ -1,6 +1,7 @@
 package com.jakewharton.mosaic.terminal
 
-@TestApi
+internal expect fun PlatformInputWriter(): PlatformInputWriter
+
 internal expect class PlatformInputWriter : AutoCloseable {
 	val input: PlatformInput
