+++ mosaic-terminal/src/jvmTest/kotlin/com/jakewharton/mosaic/terminal/PlatformInputWriter.kt
@@ -1,10 +1,29 @@
 package com.jakewharton.mosaic.terminal
 
+import com.jakewharton.mosaic.terminal.Jni.platformEventHandlerFree
+import com.jakewharton.mosaic.terminal.Jni.platformEventHandlerInit
+import com.jakewharton.mosaic.terminal.Jni.stdinWriterGetReader
+import com.jakewharton.mosaic.terminal.Jni.stdinWriterInit
 import com.jakewharton.mosaic.terminal.event.Event
 import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.channels.Channel.Factory.UNLIMITED
 
-// TODO @JvmSynthetic https://youtrack.jetbrains.com/issue/KT-24981
-internal actual class PlatformInputWriter internal constructor(
+internal actual fun PlatformInputWriter(): PlatformInputWriter {
+	val events = Channel<Event>(UNLIMITED)
+	val handlerPtr = platformEventHandlerInit(PlatformEventHandler(events))
+	if (handlerPtr != 0L) {
+		val writerPtr = stdinWriterInit(handlerPtr)
+		if (writerPtr != 0L) {
+			val readerPtr = stdinWriterGetReader(writerPtr)
+			val platformInput = PlatformInput(readerPtr, handlerPtr)
+			return PlatformInputWriter(writerPtr, events, platformInput)
+		}
+		platformEventHandlerFree(handlerPtr)
+	}
+	throw OutOfMemoryError()
+}
+
+internal actual class PlatformInputWriter(
 	private var writerPtr: Long,
 	private val events: Channel<Event>,
 	actual val input: PlatformInput,
