+++ mosaic-terminal/src/nativeMain/kotlin/com/jakewharton/mosaic/terminal/PlatformInput.kt
@@ -1,10 +1,15 @@
 package com.jakewharton.mosaic.terminal
 
+import kotlinx.cinterop.COpaquePointer
 import kotlinx.cinterop.CPointer
+import kotlinx.cinterop.NativePlacement
 import kotlinx.cinterop.StableRef
 import kotlinx.cinterop.addressOf
+import kotlinx.cinterop.alloc
+import kotlinx.cinterop.asStableRef
 import kotlinx.cinterop.free
 import kotlinx.cinterop.nativeHeap
+import kotlinx.cinterop.staticCFunction
 import kotlinx.cinterop.useContents
 import kotlinx.cinterop.usePinned
 
@@ -51,4 +56,34 @@ internal actual class PlatformInput internal constructor(
 			Tty.throwError(error)
 		}
 	}
+}
+
+internal fun StableRef<PlatformEventHandler>.toNativeAllocationIn(memory: NativePlacement): platformEventHandler {
+	return memory.alloc<platformEventHandler> {
+		opaque = asCPointer()
+		onFocus = staticCFunction(::onFocusCallback)
+		onKey = staticCFunction(::onKeyCallback)
+		onMouse = staticCFunction(::onMouseCallback)
+		onResize = staticCFunction(::onResizeCallback)
+	}
+}
+
+private fun onFocusCallback(opaque: COpaquePointer?, focused: Boolean) {
+	val handler = opaque!!.asStableRef<PlatformEventHandler>().get()
+	handler.onFocus(focused)
+}
+
+private fun onKeyCallback(opaque: COpaquePointer?) {
+	val handler = opaque!!.asStableRef<PlatformEventHandler>().get()
+	handler.onKey()
+}
+
+private fun onMouseCallback(opaque: COpaquePointer?) {
+	val handler = opaque!!.asStableRef<PlatformEventHandler>().get()
+	handler.onMouse()
+}
+
+private fun onResizeCallback(opaque: COpaquePointer?, columns: Int, rows: Int, width: Int, height: Int) {
+	val handler = opaque!!.asStableRef<PlatformEventHandler>().get()
+	handler.onResize(columns, rows, width, height)
 }
