+++ mosaic-terminal/src/nativeMain/kotlin/com/jakewharton/mosaic/terminal/Tty.kt
@@ -1,15 +1,11 @@
 package com.jakewharton.mosaic.terminal
 
 import com.jakewharton.mosaic.terminal.event.Event
-import kotlinx.cinterop.COpaquePointer
 import kotlinx.cinterop.CPointer
 import kotlinx.cinterop.StableRef
-import kotlinx.cinterop.alloc
-import kotlinx.cinterop.asStableRef
 import kotlinx.cinterop.free
 import kotlinx.cinterop.nativeHeap
 import kotlinx.cinterop.ptr
-import kotlinx.cinterop.staticCFunction
 import kotlinx.cinterop.useContents
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.Channel.Factory.UNLIMITED
@@ -38,13 +34,7 @@ public actual object Tty {
 
 		val handler = PlatformEventHandler(events)
 		val handlerRef = StableRef.create(handler)
-		val handlerPtr = nativeHeap.alloc<platformEventHandler> {
-			opaque = handlerRef.asCPointer()
-			onFocus = staticCFunction(::onFocusCallback)
-			onKey = staticCFunction(::onKeyCallback)
-			onMouse = staticCFunction(::onMouseCallback)
-			onResize = staticCFunction(::onResizeCallback)
-		}.ptr
+		val handlerPtr = handlerRef.toNativeAllocationIn(nativeHeap).ptr
 
 		val readerPtr = stdinReader_init(handlerPtr).useContents {
 			reader?.let { return@useContents it }
@@ -60,56 +50,7 @@ public actual object Tty {
 		return TerminalReader(reader, events, emitDebugEvents)
 	}
 
-	internal actual fun platformInputWriter(): PlatformInputWriter {
-		val events = Channel<Event>(UNLIMITED)
-
-		// TODO Fix all this duplication, ownership
-		val handler = PlatformEventHandler(events)
-		val handlerRef = StableRef.create(handler)
-		val handlerPtr = nativeHeap.alloc<platformEventHandler> {
-			opaque = handlerRef.asCPointer()
-			onFocus = staticCFunction(::onFocusCallback)
-			onKey = staticCFunction(::onKeyCallback)
-			onMouse = staticCFunction(::onMouseCallback)
-			onResize = staticCFunction(::onResizeCallback)
-		}.ptr
-
-		val writerPtr = stdinWriter_init(handlerPtr).useContents {
-			writer?.let { return@useContents it }
-
-			nativeHeap.free(handlerPtr)
-			handlerRef.dispose()
-
-			check(error == 0U) { "Unable to create stdin writer: $error" }
-			throw OutOfMemoryError()
-		}
-
-		val readerPtr = stdinWriter_getReader(writerPtr)!!
-		val platformInput = PlatformInput(readerPtr, handlerPtr, handlerRef)
-		return PlatformInputWriter(writerPtr, events, platformInput)
-	}
-
 	internal fun throwError(error: UInt): Nothing {
 		throw RuntimeException(error.toString())
 	}
-}
-
-private fun onFocusCallback(opaque: COpaquePointer?, focused: Boolean) {
-	val handler = opaque!!.asStableRef<PlatformEventHandler>().get()
-	handler.onFocus(focused)
-}
-
-private fun onKeyCallback(opaque: COpaquePointer?) {
-	val handler = opaque!!.asStableRef<PlatformEventHandler>().get()
-	handler.onKey()
-}
-
-private fun onMouseCallback(opaque: COpaquePointer?) {
-	val handler = opaque!!.asStableRef<PlatformEventHandler>().get()
-	handler.onMouse()
-}
-
-private fun onResizeCallback(opaque: COpaquePointer?, columns: Int, rows: Int, width: Int, height: Int) {
-	val handler = opaque!!.asStableRef<PlatformEventHandler>().get()
-	handler.onResize(columns, rows, width, height)
 }
